requirements.txt
mysql-connector-python==8.0.27
tkcalendar==1.6.1
